+++ pdfminer/pdfdocument.py
@@ -44,7 +44,11 @@ class PDFPasswordIncorrect(PDFEncryptionError):
     pass
 
 
-class PDFTextExtractionNotAllowed(PDFEncryptionError):
+class PDFTextExtractionNotAllowedWarning(UserWarning):
+    pass
+
+
+class PDFTextExtractionNotAllowedError(PDFEncryptionError):
     pass