+++ tests/test_tools_pdf2txt.py
@@ -60,12 +60,17 @@ class TestPdf2Txt():
     def test_contrib_2b(self):
         run('contrib/2b.pdf', '-A -t xml')
 
+    def test_contrib_issue_350(self):
+        """Regression test for
+        https://github.com/pdfminer/pdfminer.six/issues/350"""
+        run('contrib/issue-00352-asw-oct96-p41.pdf')
+
     def test_scancode_patchelf(self):
         """Regression test for # https://github.com/euske/pdfminer/issues/96"""
         run('scancode/patchelf.pdf')
 
     def test_contrib_hash_two_complement(self):
-        """Check that unsigned integer is added correctly to encryption hash.
+        """Check that unsigned integer is added correctly to encryption hash.et
 
         See https://github.com/pdfminer/pdfminer.six/issues/186
         """